kis_auto_trader/requirements.txt
# ═══════════════════════════════════════════════════════════════
# KIS Auto Trader - Python Dependencies
# ═══════════════════════════════════════════════════════════════

# HTTP 요청
requests>=2.31.0

# 환경변수 관리
python-dotenv>=1.0.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 스케줄링
schedule>=1.2.0

# 테스트
pytest>=8.0.0
